+++ doc_usercustomresource_test.go
@@ -54,7 +54,9 @@ func ExampleLambdaAWSInfo_RequireCustomResource() {
 		resourceMetadata map[string]interface{},
 		S3Bucket string,
 		S3Key string,
-		template *gocf.Template,
+		buildID string,
+		cfTemplate *gocf.Template,
+		context map[string]interface{},
 		logger *logrus.Logger) error {
 
 		// Pass CustomResource outputs to the λ function